netconsole: resume previously deactivated target
Attempt to resume a previously deactivated target when the associated
interface comes back (NETDEV_UP event is received) by calling
__netpoll_setup on the device.

For targets that were initally setup by mac address, their address is
also compared with the interface address.

__netpoll_setup assumes RTNL is held (which is guaranteed to be the case
when handling the event) and requires setting up the reference in case
of success.

Target transitions to STATE_DISABLED in case of failures resuming it to
avoid retrying the same target indefinitely.

Signed-off-by: Andre Carvalho <asantostc@gmail.com>